Move Reports Table to its own component
